+++ runtimes/common/src/lib.rs
@@ -44,7 +44,7 @@ use sp_runtime::{
 };
 use sp_std::marker::PhantomData;
 
-pub mod asset_swap;
+pub mod asset_switch;
 pub mod assets;
 pub mod authorization;
 pub mod constants;
